+++ isodata/isone.py
@@ -1,12 +1,15 @@
 import io
+import math
 import re
+from tkinter import E
 from urllib import request
 
 import pandas as pd
 import requests
 
 import isodata
-from isodata.base import FuelMix, ISOBase
+from isodata import utils
+from isodata.base import FuelMix, ISOBase, Markets
 
 
 class ISONE(ISOBase):
@@ -14,13 +17,40 @@ class ISONE(ISOBase):
     iso_id = "isone"
     default_timezone = "US/Eastern"
 
+    REAL_TIME_5_MIN = Markets.REAL_TIME_5_MIN
+    REAL_TIME_HOURLY = Markets.REAL_TIME_HOURLY
+    DAY_AHEAD_HOURLY = Markets.DAY_AHEAD_HOURLY
+
+    hubs = {"H.INTERNAL_HUB": 4000}
+    zones = {
+        ".Z.MAINE": 4001,
+        ".Z.NEWHAMPSHIRE": 4002,
+        ".Z.VERMONT": 4003,
+        ".Z.CONNECTICUT": 4004,
+        ".Z.RHODEISLAND": 4005,
+        ".Z.SEMASS": 4006,
+        ".Z.WCMASS": 4007,
+        ".Z.NEMASSBOST": 4008,
+    }
+    interfaces = {
+        ".I.SALBRYNB345": 4010,
+        ".I.ROSETON 345": 4011,
+        ".I.HQ_P1_P2345": 4012,
+        ".I.HQHIGATE120": 4013,
+        ".I.SHOREHAM138": 4014,
+        ".I.NRTHPORT138": 4017,
+    }
+
     def get_latest_fuel_mix(self):
         r = requests.post(
             "https://www.iso-ne.com/ws/wsclient",
             data={"_nstmp_requestType": "fuelmix"},
         ).json()
         mix_df = pd.DataFrame(r[0]["data"]["GenFuelMixes"]["GenFuelMix"])
-        time = pd.Timestamp(mix_df["BeginDate"].max(), tz=self.default_timezone)
+        time = pd.Timestamp(
+            mix_df["BeginDate"].max(),
+            tz=self.default_timezone,
+        )
 
         # todo has marginal flag
         mix_dict = mix_df.set_index("FuelCategory")["GenMw"].to_dict()
@@ -46,29 +76,7 @@ class ISONE(ISOBase):
             "%Y%m%d",
         )
 
-        with requests.Session() as s:
-            # in testing, never takes more than 2 attempts
-            attempt = 0
-            while attempt < 3:
-                # make first get request to get cookies set
-                r1 = s.get(
-                    "https://www.iso-ne.com/isoexpress/web/reports/operations/-/tree/gen-fuel-mix",
-                )
-
-                r2 = s.get(url)
-
-                if r2.status_code == 200:
-                    break
-
-                print("Attempt {} failed. Retrying...".format(attempt + 1))
-                attempt += 1
-
-            df = pd.read_csv(
-                io.StringIO(r2.content.decode("utf8")),
-                skiprows=[0, 1, 2, 3, 5],
-                skipfooter=1,
-                engine="python",
-            )
+        df = _make_request(url, skiprows=[0, 1, 2, 3, 5])
 
         df["Date"] = pd.to_datetime(df["Date"] + " " + df["Time"]).dt.tz_localize(
             self.default_timezone,
@@ -115,7 +123,10 @@ class ISONE(ISOBase):
             "_nstmp_inclBtmPv": True,
         }
 
-        r = requests.post("https://www.iso-ne.com/ws/wsclient", data=data).json()
+        r = requests.post(
+            "https://www.iso-ne.com/ws/wsclient",
+            data=data,
+        ).json()
 
         data = pd.DataFrame(r[0]["data"]["actual"])
 
@@ -145,7 +156,186 @@ class ISONE(ISOBase):
         """Returns supply at a previous date in MW"""
         return self._supply_from_fuel_mix(date)
 
+    def get_latest_lmp(self, market: str, nodes: list):
+        """
+        Find Node ID mapping: https://www.iso-ne.com/markets-operations/settlements/pricing-node-tables/
+        """
+        # todo optimize to read latest csv
+        if market == self.REAL_TIME_5_MIN:
+            url = "https://www.iso-ne.com/transform/csv/fiveminlmp/current?type=prelim"
+            data = _make_request(url, skiprows=[0, 1, 2, 4])
+        elif market == self.REAL_TIME_HOURLY:
+            url = "https://www.iso-ne.com/transform/csv/hourlylmp/current?type=prelim&market=rt"
+            data = _make_request(url, skiprows=[0, 1, 2, 4])
 
-# daily historical fuel mix
-# https://www.iso-ne.com/static-assets/documents/2022/01/2022_daygenbyfuel.xlsx
-# a bunch more here: https://www.iso-ne.com/isoexpress/web/reports/operations/-/tree/daily-gen-fuel-type
+            # todo does this handle single digital hours?
+            data["Local Time"] = (
+                data["Local Date"]
+                + " "
+                + data["Local Time"].astype(str).str.zfill(2)
+                + ":00"
+            )
+
+        else:
+            raise RuntimeError("Unsupported market")
+
+        data = _process_lmp(data, market, self.default_timezone, nodes)
+        return data
+
+    def get_lmp_today(self, market: str, nodes: list):
+        "Get lmp for today in 5 minute intervals"
+        return self._today_from_historical(self.get_historical_lmp, market, nodes)
+
+    def get_lmp_yesterday(self, market: str, nodes: list):
+        "Get lmp for yesterday in 5 minute intervals"
+        return self._yesterday_from_historical(self.get_historical_lmp, market, nodes)
+
+    def get_historical_lmp(self, date, market: str, nodes: list):
+        """Find Node ID mapping: https://www.iso-ne.com/markets-operations/settlements/pricing-node-tables/"""
+        date = isodata.utils._handle_date(date)
+        date_str = date.strftime("%Y%m%d")
+
+        now = pd.Timestamp.now(tz=self.default_timezone)
+
+        if market == self.REAL_TIME_5_MIN:
+            # todo handle intervals for current day
+            intervals = ["00-04", "04-08", "08-12", "12-16", "16-20", "20-24"]
+
+            # optimze for current day
+            if now.date() == date.date():
+                hour = now.hour
+                # select completed 4 hour intervals based on current hour
+                intervals = intervals[: math.ceil((hour + 1) / 4) - 1]
+
+            dfs = []
+            for interval in intervals:
+                print("Loading interval {}".format(interval))
+                u = f"https://www.iso-ne.com/static-transform/csv/histRpts/5min-rt-prelim/lmp_5min_{date_str}_{interval}.csv"
+                dfs.append(
+                    pd.read_csv(
+                        u,
+                        skiprows=[0, 1, 2, 3, 5],
+                        skipfooter=1,
+                        engine="python",
+                    ),
+                )
+
+            data = pd.concat(dfs)
+
+            data["Local Time"] = (
+                date.strftime(
+                    "%Y-%m-%d",
+                )
+                + " "
+                + data["Local Time"]
+            )
+
+            # add current interval
+            if now.date() == date.date():
+                url = "https://www.iso-ne.com/transform/csv/fiveminlmp/currentrollinginterval"
+                print("Loading current interval")
+                data_current = _make_request(url, skiprows=[0, 1, 2, 4])
+
+                data_current = data_current[
+                    data_current["Local Time"] > data["Local Time"].max()
+                ]
+
+                data = pd.concat([data, data_current])
+
+        elif market == self.REAL_TIME_HOURLY:
+            if date.date() < now.date():
+                url = f"https://www.iso-ne.com/static-transform/csv/histRpts/rt-lmp/lmp_rt_prelim_{date_str}.csv"
+                data = _make_request(url, skiprows=[0, 1, 2, 3, 5])
+                # todo document hour starting vs ending
+                data["Local Time"] = (
+                    data["Date"]
+                    + " "
+                    + (data["Hour Ending"] - 1).astype(str).str.zfill(2)
+                    + ":00"
+                )
+            else:
+                raise RuntimeError("Today not support for hourly lmp")
+
+        elif market == self.DAY_AHEAD_HOURLY:
+            url = f"https://www.iso-ne.com/static-transform/csv/histRpts/da-lmp/WW_DALMP_ISO_{date_str}.csv"
+            data = _make_request(url, skiprows=[0, 1, 2, 3, 5])
+            # todo document hour starting vs ending
+            data["Local Time"] = (
+                data["Date"]
+                + " "
+                + (data["Hour Ending"] - 1).astype(str).str.zfill(2)
+                + ":00"
+            )
+        else:
+            raise RuntimeError("Unsupported market")
+
+        data = _process_lmp(data, market, self.default_timezone, nodes)
+
+        return data
+
+        # daily historical fuel mix
+        # https://www.iso-ne.com/static-assets/documents/2022/01/2022_daygenbyfuel.xlsx
+        # a bunch more here: https://www.iso-ne.com/isoexpress/web/reports/operations/-/tree/daily-gen-fuel-type
+
+
+def _make_request(url, skiprows):
+    with requests.Session() as s:
+        # in testing, never takes more than 2 attempts
+        attempt = 0
+        while attempt < 3:
+            # make first get request to get cookies set
+            r1 = s.get(
+                "https://www.iso-ne.com/isoexpress/web/reports/operations/-/tree/gen-fuel-mix",
+            )
+
+            r2 = s.get(url)
+
+            if r2.status_code == 200:
+                break
+
+            print("Attempt {} failed. Retrying...".format(attempt + 1))
+            attempt += 1
+
+        df = pd.read_csv(
+            io.StringIO(r2.content.decode("utf8")),
+            skiprows=skiprows,
+            skipfooter=1,
+            engine="python",
+        )
+        return df
+
+
+def _process_lmp(data, market, timezone, nodes):
+    # todo handle location types
+    rename = {
+        "Location ID": "Node",
+        "Location": "Node",
+        "Local Time": "Time",
+        "Locational Marginal Price": "LMP",
+        "LMP": "LMP",
+        "Energy Component": "Energy",
+        "Congestion Component": "Congestion",
+        "Loss Component": "Loss",
+        "Marginal Loss Component": "Loss",
+    }
+
+    data.rename(columns=rename, inplace=True)
+
+    data["Market"] = market
+
+    data["Time"] = pd.to_datetime(data["Time"]).dt.tz_localize(timezone)
+
+    data = data[
+        [
+            "Time",
+            "Market",
+            "Node",
+            "LMP",
+            "Energy",
+            "Congestion",
+            "Loss",
+        ]
+    ]
+
+    data = utils.filter_lmp_nodes(data, nodes)
+    return data